tirei o padrao singleton das materias lascou a porra toda

+++ Quadro/FotosViewController.h
@@ -10,6 +10,6 @@
 
 @interface FotosViewController : UIViewController <UICollectionViewDataSource, UICollectionViewDelegate, UIActionSheetDelegate, UITextViewDelegate>
 
-@property (nonatomic) NSInteger posicaoMateria, posicaoAssunto;
+@property (nonatomic) int posicaoMateria, posicaoAssunto;
 
 @end
